refactor: separate logic of import match in extra file #350
