+++ src/helper/helper.ts
@@ -3,7 +3,7 @@ import path from "node:path";
 import { type QueryCapture, type Tree } from "tree-sitter";
 import { type Configuration } from "../parser/configuration.js";
 import { NodeTypeQueryStatement } from "../parser/queries/query-statements.js";
-import { type ImportCapture } from "../parser/resolver/call-expressions/abstract-collector.js";
+import { type ImportCapture } from "../parser/resolver/call-expressions/import-match.js";
 import { type NodeTypeCategory, type NodeTypeConfig } from "./model.js";
 
 /**
